Put Home Page classes to proper places and renamed HomeFragment to HomeOverviewFragment
