+++ app/src/main/java/com/tari/android/wallet/data/TransactionRepository.kt
@@ -1,4 +1,4 @@
-package com.tari.android.wallet.ui.fragment.tx
+package com.tari.android.wallet.data
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MediatorLiveData
