+++ app/src/main/java/com/tari/android/wallet/ui/component/balanceController/BalanceViewController.kt
@@ -30,7 +30,7 @@
  * OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.tari.android.wallet.ui.fragment.tx.ui.balanceController
+package com.tari.android.wallet.ui.component.balanceController
 
 import android.animation.Animator
 import android.animation.AnimatorListenerAdapter
