+++ app/src/main/java/com/tari/android/wallet/ui/fragment/auth/AuthViewModel.kt
@@ -11,7 +11,7 @@ import com.tari.android.wallet.ui.dialog.modular.modules.body.BodyModule
 import com.tari.android.wallet.ui.dialog.modular.modules.button.ButtonModule
 import com.tari.android.wallet.ui.dialog.modular.modules.button.ButtonStyle
 import com.tari.android.wallet.ui.dialog.modular.modules.head.HeadModule
-import com.tari.android.wallet.ui.fragment.home.navigation.Navigation
+import com.tari.android.wallet.navigation.Navigation
 import javax.inject.Inject
 
 class AuthViewModel : CommonViewModel() {
