+++ app/src/main/java/com/tari/android/wallet/ui/fragment/contactBook/contactSelection/ContactSelectionViewModel.kt
@@ -35,7 +35,7 @@ import com.tari.android.wallet.ui.fragment.contactBook.data.contacts.ContactDto
 import com.tari.android.wallet.ui.fragment.contactBook.data.contacts.FFIContactInfo
 import com.tari.android.wallet.ui.fragment.contactBook.data.contacts.splitAlias
 import com.tari.android.wallet.ui.fragment.contactBook.root.ShareViewModel
-import com.tari.android.wallet.ui.fragment.home.navigation.Navigation
+import com.tari.android.wallet.navigation.Navigation
 import com.tari.android.wallet.util.Constants
 import com.tari.android.wallet.util.EmojiId
 import com.tari.android.wallet.util.EmojiUtil.Companion.getGraphemeLength
