+++ app/src/main/java/com/tari/android/wallet/ui/fragment/contactBook/details/ContactDetailsFragment.kt
@@ -14,7 +14,7 @@ import com.tari.android.wallet.ui.component.tari.toolbar.TariToolbarActionArg
 import com.tari.android.wallet.ui.fragment.contactBook.data.ContactAction
 import com.tari.android.wallet.ui.fragment.contactBook.data.contacts.ContactDto
 import com.tari.android.wallet.ui.fragment.contactBook.details.adapter.ContactDetailsAdapter
-import com.tari.android.wallet.ui.fragment.home.navigation.TariNavigator.Companion.PARAMETER_CONTACT
+import com.tari.android.wallet.navigation.TariNavigator.Companion.PARAMETER_CONTACT
 
 class ContactDetailsFragment : CommonFragment<FragmentContactsDetailsBinding, ContactDetailsViewModel>() {
