+++ app/src/main/java/com/tari/android/wallet/ui/fragment/send/addNote/AddNoteViewModel.kt
@@ -3,7 +3,7 @@ package com.tari.android.wallet.ui.fragment.send.addNote
 import com.tari.android.wallet.model.TariWalletAddress
 import com.tari.android.wallet.network.NetworkConnectionStateHandler
 import com.tari.android.wallet.ui.common.CommonViewModel
-import com.tari.android.wallet.ui.fragment.home.navigation.Navigation
+import com.tari.android.wallet.navigation.Navigation
 import com.tari.android.wallet.ui.fragment.send.common.TransactionData
 import javax.inject.Inject
